functions/requirements.txt
# Firebase Functions dependencies
functions-framework==3.*
flask==2.3.*
firebase-admin==6.*
google-cloud-functions==1.*
google-generativeai==0.8.*
python-dotenv==1.0.*
requests==2.31.*
